ms2/requirements.txt
kafka-python
faker
python-dotenv
pandas
sqlalchemy

psycopg2-binary